Change how examples are presented by help generators
Each entry is now simply presented as-is rather than trying to convert
into a list format, this was done because outputting it as previously
implemented was particularly fragile when converting to target formats
and required that users include markup in their annotations which is
undesireable.

Users are of course free to extend/implement a custom help generator
that presents examples differently if desired.