Initial work on command abbreviation (#7)
This commit adds the initial support for command and group abbreviation.
This allows the option for enabling support for command abbreviation
which allows users to not type out the full name of a group/command
provided said group/command is not an ambigious name.

Adds basic test cases for this but not yet coverage of the awkward cases