Fix a bug that could occur when overriding options
When an option overrode the type of its parent option with a narrowing
type conversion this was permitted but occassionally we could report the
wrong type as the effective type of the option.  Now ensures that the
narrowest type is the first in the list of accessors and so always
reported as the type of the option.