Improve some help generators (#12)
Some help generators did not actually allow you to configure whether
they included hidden commands and options even if they actually
supported and honoured that option.  This commit fixes this for
CliCommandUsageGenerator and starts adding new `Help.help()` overloads
that allow the inclusion of hidden commands/options to be specified.