Extract user alias config into own class
Previously while user defined aliases could be used they were loaded as
soon as they were specified and as such it was not possible to know if
aliases originated from user configuration files after the fact.  This
commit splits the information and their loading out into a separate
class UserAliasesSource the intention of which is that it can be used by
other code to inspect information about user aliases and present that as
they see fit.