Convert `RestrictionRegistry` to using `ServiceLoader` (#33)
This commit converts the `RestrictionRegistry` to using `ServiceLoader`
to automatically discover restriction annotations and their associated
factories.  This required a breaking change to the
`OptionRestrictionFactory` and `ArgumentsRestrictionFactory` interfaces
to allow them to return the list of annotations they support.

New files are added to `META-INF/services/` which specify the built-in
set of restriction factories.

Updated the Change Log and Migration Guide to reflect this