Fix for incorrect indendation in Man options section (#34)
This commit fixes a bug in how the Man command output decides to indent
the arguments portion of the options section of help that could cause it
to indent it incorrectly when there were both visible and hidden options
present.

It also fixes another related bug that could cause us to fail generating
help if only hidden options are present.

Various unit tests for these cases are added to the test suite