Change Restriction APIs to support Partial Restrictions (#36)
This commit changes the OptionRestriction and ArgumentsRestriction API
so that they can properly support the partial restrictions mechanism.
The existing postValidate() method is replaced with a finalValidate()
method while a new postValidate() method is introduced which has a
similar signature to preValidate() except that it receives the value
after type conversion.

Unit tests related to partial restrictions are extended to test that the
refactoring resolves the issue previously identified and to further test
partial restrictions.